The newest itteration of the Application
I've introduced a new class and object element in the NewTime(tk).py file. Instead of the previous procedural approach I have decided it will be best to follow along with tkinter's natural object orientation. There are a few other elements which are cleaned images and another import method that I want to introduce.